Demo AI Project
This is the full demo thus far. It has AI elements from Clive and my
projects. There are also a few changes made to the player. The
player_motor and FPSCharacterController are now written in C# instead of
javascript. The player can also shoot and kill the revenants. (and he
can be killed...) There is no current respawn.